Add new checkbox.js file for complete / pending logic
