+++ src/index.js
@@ -1,6 +1,7 @@
 import './style.css';
 import { dragStart, dragOver, drop } from './dragandrop.js';
 import { loadTasks, saveTasks, reorderTasks } from './task.js';
+import taskCompleteOrPending from './checkbox.js';
 
 const createElement = (name) => document.createElement(name);
 const createElementWithClass = (name, clazz) => {
@@ -22,26 +23,7 @@ const createCheckBoxElement = (task, checkboxId) => {
     checkboxElement.setAttribute('checked', 'checked');
   }
 
-  checkboxElement.addEventListener('change', (e) => {
-    const checkbox = e.target;
-    const taskIndex = checkbox.id.substring('check-task-'.length);
-    const descriptionElement = document.getElementById(`todo-list-task-description-${taskIndex}`);
-
-    if (checkbox.checked) {
-      descriptionElement.classList.add('strikethrough');
-    } else {
-      descriptionElement.classList.remove('strikethrough');
-    }
-
-    const tasks = loadTasks();
-    for (const i in tasks) { /* eslint-disable-line no-restricted-syntax */
-      if (tasks[i].index == taskIndex) { /* eslint-disable-line eqeqeq */
-        tasks[i].completed = !tasks[i].completed;
-      }
-    }
-
-    saveTasks(tasks);
-  });
+  checkboxElement.addEventListener('change', taskCompleteOrPending);
   return checkboxElement;
 };
 const createDescriptionDiv = (task) => {
